backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
langsmith
python-dotenv
